Implemented GameOver/Win Screen + Some Pause Menu fixes
Added game over and win screen feature that shows a retry button, quit button and dynamic text when the game's objective is cleared or when all player characters are defeated. Retry reloads the scene and quit returns to the main menu. Also added a background to other parts of the pause menu to prevent the player from being able to click on the grid in the placeholder help and options menus.